backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
playwright
pywebview
sqlalchemy
psutil
supabase
python-dotenv
thefuzz
python-Levenshtein
